requirements.txt
Flask==2.3.3
pymongo==4.5.0
python-decouple==3.8
gunicorn==21.2.0
